+++ chaosaws/__init__.py
@@ -252,8 +252,9 @@ def load_exported_activities() -> List[DiscoveredActivities]:
     activities.extend(discover_actions("chaosaws.emr.actions"))
     activities.extend(discover_probes("chaosaws.emr.probes"))
     activities.extend(discover_actions("chaosaws.route53.actions"))
-    activities.extend(discover_probes('chaosaws.route53.probes'))
-    activities.extend(discover_actions('chaosaws.ssm.actions'))
+    activities.extend(discover_probes("chaosaws.route53.probes"))
+    activities.extend(discover_probes("chaosaws.ssm.actions"))
+    activities.extend(discover_actions("chaosaws.fis.actions"))
     activities.extend(discover_probes('chaosaws.s3.probes'))
     activities.extend(discover_actions('chaosaws.s3.actions'))
     return activities